feat: add try catch and updatePassword route

+++ src/user/user.controller.ts
@@ -3,6 +3,7 @@ import {
     Controller,
     Delete,
     Get,
+    HttpException,
     HttpStatus,
     Param,
     Patch,
@@ -11,6 +12,7 @@ import {
 } from '@nestjs/common';
 import { Response } from 'express';
 import { CreateUserDto } from './dto/create-user.dto';
+import { UpdatePasswordDto } from './dto/update-password-dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserService } from './user.service';
 
@@ -20,24 +22,41 @@ export class UserController {
 
     @Post()
     async create(@Body() createUserDto: CreateUserDto, @Res() res: Response) {
-        const checkUser = await this.usersService.checkUniqueUser(
-            createUserDto.email,
-            createUserDto.username,
-        );
+        try {
+            const checkUser = await this.usersService.checkUniqueUser(
+                createUserDto.email,
+                createUserDto.username,
+            );
 
-        if (!checkUser) {
-            return res.status(HttpStatus.CONFLICT).json();
+            if (!checkUser) {
+                throw new HttpException(
+                    'User already exists.',
+                    HttpStatus.CONFLICT,
+                );
+            }
+
+            if (
+                !createUserDto.displayname ||
+                createUserDto.displayname === ''
+            ) {
+                createUserDto.displayname = createUserDto.username;
+            }
+
+            const user = await this.usersService.create(createUserDto);
+
+            const response = {
+                uuid: user.uuid,
+                username: user.username,
+                displayname: user.displayname,
+                bio: user.bio,
+                email: user.email,
+                createdAt: user.createdAt,
+            };
+
+            return res.status(HttpStatus.CREATED).json(response);
+        } catch (error) {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json(error);
         }
-
-        const user = await this.usersService.create(createUserDto);
-
-        const response = {
-            uuid: user.uuid,
-            username: user.username,
-            email: user.email,
-        };
-
-        return res.status(HttpStatus.CREATED).json(response);
     }
 
     @Get()
@@ -50,6 +69,10 @@ export class UserController {
                     uuid: user.uuid,
                     username: user.username,
                     email: user.email,
+                    displayname: user.displayname,
+                    bio: user.bio,
+                    createdAt: user.createdAt,
+                    updatedAt: user.updatedAt,
                 };
             });
 
@@ -63,6 +86,12 @@ export class UserController {
     async findOne(@Param('username') username: string, @Res() res: Response) {
         try {
             const user = await this.usersService.findOne(username);
+            if (!user) {
+                throw new HttpException(
+                    'User not found.',
+                    HttpStatus.NOT_FOUND,
+                );
+            }
 
             return res.status(HttpStatus.OK).json({
                 uuid: user.uuid,
@@ -81,8 +110,47 @@ export class UserController {
         @Res() res: Response,
     ) {
         try {
+            const user = await this.usersService.findOne(username);
+            if (!user) {
+                throw new HttpException(
+                    'User not found.',
+                    HttpStatus.NOT_FOUND,
+                );
+            }
+
             await this.usersService.update(username, updateUserDto);
             return res.status(HttpStatus.OK).json('User updated successfully.');
+        } catch (error) {
+            return res.status(HttpStatus.CONFLICT).json(error);
+        }
+    }
+
+    @Patch('/updatePassword')
+    async updatePassword(
+        @Body() updatePasswordDto: UpdatePasswordDto,
+        @Res() res: Response,
+    ) {
+        try {
+            const user = await this.usersService.findOne(
+                updatePasswordDto.email,
+            );
+
+            if (!user) {
+                throw new HttpException(
+                    'User not found.',
+                    HttpStatus.NOT_FOUND,
+                );
+            }
+
+            await this.usersService.updatePassword(
+                updatePasswordDto.email,
+                updatePasswordDto.password,
+                updatePasswordDto.newPassword,
+            );
+
+            return res
+                .status(HttpStatus.NO_CONTENT)
+                .json('Password updated successfully.');
         } catch (error) {
             return res.status(HttpStatus.CONFLICT).json(error);
         }
@@ -91,6 +159,14 @@ export class UserController {
     @Delete(':username')
     async remove(@Param('username') username: string, @Res() res: Response) {
         try {
+            const user = await this.usersService.findOne(username);
+            if (!user) {
+                throw new HttpException(
+                    'User not found.',
+                    HttpStatus.NOT_FOUND,
+                );
+            }
+
             await this.usersService.remove(username);
             return res.status(HttpStatus.OK).json('User deleted successfully.');
         } catch (error) {
